manual_generator/requirements.txt
streamlit
google-generativeai
opencv-python-headless
python-docx